[#657] Renamed _mesh.py to _class0_Plasma2D.py
