+++ tofu/data/__init__.py
@@ -5,5 +5,5 @@ Provide data handling class and methods (storing, processing, plotting...)
 from ._core import *
 from ._DataCollection_class1_interactivity import *
 from ._spectrallines_class import *
-from ._mesh import *
+from ._class0_Plasma2D import *
 from ._inversions_checks import get_available_inversions_algo
